[UPDATE] middleware combined
